components tested

+++ tests/test_plasma.py
@@ -1,64 +1,70 @@
 #
 #    ICRAR - International Centre for Radio Astronomy Research
-#    (c) UWA - The University of Western Australia, 2015
+#    (c) UWA - The University of Western Australia, 2021
 #    Copyright by UWA (in the framework of the ICRAR)
 #    All rights reserved
 #
-#    This library is free software; you can redistribute it and/or
-#    modify it under the terms of the GNU Lesser General Public
-#    License as published by the Free Software Foundation; either
-#    version 2.1 of the License, or (at your option) any later version.
+#     This program is free software: you can redistribute it and/or modify
+#     it under the terms of the GNU General Public License as published by
+#     the Free Software Foundation, either version 3 of the License, or
+#     (at your option) any later version.
 #
-#    This library is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-#    Lesser General Public License for more details.
+#     This program is distributed in the hope that it will be useful,
+#     but WITHOUT ANY WARRANTY; without even the implied warranty of
+#     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#     GNU General Public License for more details.
 #
-#    You should have received a copy of the GNU Lesser General Public
-#    License along with this library; if not, write to the Free Software
-#    Foundation, Inc., 59 Temple Place, Suite 330, Boston,
-#    MA 02111-1307  USA
+#     You should have received a copy of the GNU General Public License
+#     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #
-import unittest
+import os
+import time
+import sys
 import logging
 import tarfile
 import binascii
 import shutil
+import subprocess
+import unittest
+from pathlib import Path
+
+import numpy as np
+import pyarrow.plasma as plasma
+from casacore import tables
 
 from dlg.drop import FileDROP, PlasmaDROP, InMemoryDROP
 from dlg import droputils
+from dlg.numpydroputils import NumpyDropUtils
 
-cbf_unavailable = True
-try:
-    from cbf_sdp import msutils
+from cbf_sdp.ms_asserter import MSAsserter
+from dlg_casacore_components.plasma import MSPlasmaWriter, MSPlasmaReader
+from dlg_casacore_components.cbf_sdp import MSStreamingPlasmaConsumer, MSStreamingPlasmaProducer
 
-    cbf_unavailable = False
-except Exception as e:
-    print(e)
-    pass
+logging.basicConfig(level=logging.INFO, handlers=[logging.StreamHandler(sys.stdout)])
 
-
-import pyarrow.plasma as plasma
-from dlg.apps.plasma import MSPlasmaWriter, MSPlasmaReader
-from dlg.apps.plasma import MSStreamingPlasmaConsumer, MSStreamingPlasmaProducer
-from casacore import tables
-
-logging.basicConfig()
-
+test_ms_dir = Path(__file__).parent.absolute() / "data/test_ms.tar.gz"
 
 class CRCAppTests(unittest.TestCase):
+    def setUp(self):
+        self.store = subprocess.Popen(
+            ["plasma_store", "-m", "100000000", "-s", "/tmp/plasma"]
+        )
+
+    def tearDown(self):
+        self.store.terminate()
+
     def compare_measurement_sets(self, in_file, out_file):
-        asserter = type("asserter", (msutils.MSAsserter, unittest.TestCase), {})()
+        asserter = type("asserter", (MSAsserter, unittest.TestCase), {})()
         asserter.assert_ms_data_equal(in_file, out_file)
 
     def compare_ms(self, in_file, out_file):
         a = []
         b = []
-        with tables.table(out_file) as t1:
+        with tables.table(out_file, ack=False) as t1:
             for i in t1:
                 a.append(i["DATA"])
 
-        with tables.table(in_file) as t2:
+        with tables.table(in_file, ack=False) as t2:
             for i in t2:
                 b.append(i["DATA"])
 
@@ -66,22 +72,63 @@ class CRCAppTests(unittest.TestCase):
             comparison = j == b[i]
             self.assertEqual(comparison.all(), True)
 
-    @unittest.skipIf(cbf_unavailable, "sdp-cbf not available")
+
+    def test_plasma_client(self):
+        import pyarrow
+        import io
+        import numpy as np
+
+        client = pyarrow.plasma.connect("/tmp/plasma")
+        indata = np.ones([10,10])
+
+        # Read+Write BytesIO
+        bio = io.BytesIO()
+        np.save(bio, indata, allow_pickle=False)
+        bio.seek(0)
+
+        # Direct Access
+        outdata = np.load(bio, allow_pickle=False)
+        np.testing.assert_array_equal(indata, outdata)
+
+        # Plasma Put+Get
+        objectid = client.put(bio.getvalue())
+        outdata = np.load(io.BytesIO(client.get(objectid)))
+        np.testing.assert_array_equal(indata, outdata)
+
+        # Plasma Raw Buffer Put+Get
+        objectid = client.put_raw_buffer(bio.getbuffer())
+        [buf] = client.get_buffers([objectid])
+        outdata = np.load(io.BytesIO(buf))
+        np.testing.assert_array_equal(indata, outdata)
+
+        # Plasma Raw Buffer Create+Seal+Get
+        objectid = plasma.ObjectID(np.random.bytes(20))
+        plasma_buffer = client.create(objectid, bio.__sizeof__())
+        writer = pyarrow.FixedSizeBufferWriter(plasma_buffer)
+        writer.write(bio.getbuffer())
+        client.seal(objectid)
+        [buf] = client.get_buffers([objectid])
+        outdata = np.load(io.BytesIO(buf))
+        np.testing.assert_array_equal(indata, outdata)
+
+
     def test_plasma_stream(self):
         in_file = "/tmp/test.ms"
         out_file = "/tmp/copy.ms"
 
-        try:
-            shutil.rmtree(in_file)
-        except:
-            pass
 
-        try:
-            shutil.rmtree(out_file)
-        except:
-            pass
+        if os.path.exists(in_file):
+            if os.path.isdir(in_file):
+                shutil.rmtree(in_file)
+            else:
+                os.remove(in_file)
+        if os.path.exists(out_file):
+            if os.path.isdir(out_file):
+                shutil.rmtree(out_file)
+            else:
+                os.remove(out_file)
 
-        with tarfile.open("/daliuge/test/apps/data/test_ms.tar.gz", "r") as ref:
+        with tarfile.open(test_ms_dir, "r") as ref:
             ref.extractall("/tmp/")
 
         prod = MSStreamingPlasmaProducer("1", "1")
@@ -97,24 +144,26 @@ class CRCAppTests(unittest.TestCase):
         with droputils.DROPWaiterCtx(self, cons, 1000):
             prod.async_execute()
 
-        # self.compare_ms(in_file, out_file)
         self.compare_measurement_sets(in_file, out_file)
+        time.sleep(5)
 
-    def test_plasma(self):
+
+    def test_plasma_writer(self):
         in_file = "/tmp/test.ms"
         out_file = "/tmp/copy.ms"
 
-        try:
-            shutil.rmtree(in_file)
-        except:
-            pass
+        if os.path.exists(in_file):
+            if os.path.isdir(in_file):
+                shutil.rmtree(in_file)
+            else:
+                os.remove(in_file)
+        if os.path.exists(out_file):
+            if os.path.isdir(out_file):
+                shutil.rmtree(out_file)
+            else:
+                os.remove(out_file)
 
-        try:
-            shutil.rmtree(out_file)
-        except:
-            pass
-
-        with tarfile.open("/daliuge/test/apps/data/test_ms.tar.gz", "r") as ref:
+        with tarfile.open(test_ms_dir, "r") as ref:
             ref.extractall("/tmp/")
 
         a = FileDROP("a", "a", filepath=in_file)
@@ -131,6 +180,7 @@ class CRCAppTests(unittest.TestCase):
         # Check the MS DATA content is the same as original
         with droputils.DROPWaiterCtx(self, e, 5):
             a.setCompleted()
+        #time.sleep(5)
 
         # self.compare_ms(in_file, out_file)
         self.compare_measurement_sets(in_file, out_file)
@@ -139,4 +189,5 @@ class CRCAppTests(unittest.TestCase):
         client = plasma.connect("/tmp/plasma")
         a = c.dataURL.split("//")[1]
         a = binascii.unhexlify(a)
-        client.get(plasma.ObjectID(a))
+        client.get_buffers([plasma.ObjectID(a)])
+
